Componentizando e refatorando a página DetailsRecipes
