+++ src/pages/DetailsRecipes/componentes/CardMidiaDetailsRecipes/styles.js
@@ -1,0 +1,7 @@
+import { CardMedia, styled } from "@mui/material";
+
+export const Midias = styled(CardMedia)(({ theme }) => ({
+    height: '20rem',
+    width: '100%',
+    backgroundSize: '100%',
+}))
