+++ src/pages/DetailsRecipes/componentes/StepsDetailsRecipe/index.js
@@ -1,0 +1,10 @@
+import { Stack, styled } from "@mui/material";
+
+export const StepsContain = styled(Stack)(({ theme }) => ({
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    textAlign: 'left',
+    width: '90%',
+    bgcolor: '',
+}))
